rgw: split simplefile into several files
Instead of managing a monolithic file, lets try doing this in smaller
units.
