+++ src/rgw/CMakeLists.txt
@@ -175,7 +175,15 @@ if(WITH_RADOSGW_DBSTORE)
   list(APPEND librgw_common_srcs rgw_sal_dbstore.cc)
 endif()
 if(WITH_RADOSGW_SIMPLEFILE)
-  list(APPEND librgw_common_srcs rgw_sal_simplefile.cc)
+  list(APPEND librgw_common_srcs
+    rgw_sal_simplefile.cc
+    store/simplefile/bucket.cc
+    store/simplefile/object.cc
+    store/simplefile/user.cc
+    store/simplefile/zone.cc
+    store/simplefile/sfs_bucket.cc
+    store/simplefile/sfs_paths.cc
+    store/simplefile/sfs_user.cc)
 endif()
 if(WITH_RADOSGW_MOTR)
   list(APPEND librgw_common_srcs rgw_sal_motr.cc)
