+++ src/rgw/rgw_sal_simplefile.cc
@@ -36,660 +36,6 @@ using namespace std;
 
 namespace rgw::sal {
 
-// Zone {{{
-
-const RGWZoneGroup &SimpleFileZone::get_zonegroup() { return *zonegroup; }
-
-int SimpleFileZone::get_zonegroup(const std::string &id, RGWZoneGroup &zg) {
-  zg = *zonegroup;
-  return 0;
-}
-
-const RGWZoneParams &SimpleFileZone::get_params() { return *zone_params; }
-
-const rgw_zone_id &SimpleFileZone::get_id() { return cur_zone_id; }
-
-const RGWRealm &SimpleFileZone::get_realm() { return *realm; }
-
-const std::string &SimpleFileZone::get_name() const {
-  return zone_params->get_name();
-}
-
-bool SimpleFileZone::is_writeable() { return true; }
-
-bool SimpleFileZone::get_redirect_endpoint(std::string *endpoint) {
-  return false;
-}
-
-bool SimpleFileZone::has_zonegroup_api(const std::string &api) const {
-  return false;
-}
-
-const std::string &SimpleFileZone::get_current_period_id() {
-  return current_period->get_id();
-}
-
-SimpleFileZone::SimpleFileZone(SimpleFileStore *_store) : store(_store) {
-  realm = new RGWRealm();
-  zonegroup = new RGWZoneGroup();
-  zone_public_config = new RGWZone();
-  zone_params = new RGWZoneParams();
-  current_period = new RGWPeriod();
-  cur_zone_id = rgw_zone_id(zone_params->get_id());
-  RGWZonePlacementInfo info;
-  RGWZoneStorageClasses sc;
-  sc.set_storage_class("STANDARD", nullptr, nullptr);
-  info.storage_classes = sc;
-  zone_params->placement_pools["default"] = info;
-}
-// }}}
-
-// Object {{{
-
-SimpleFileObject::SimpleFileReadOp::SimpleFileReadOp(SimpleFileObject *_source,
-                                                     RGWObjectCtx *_rctx)
-    : source(_source), rctx(_rctx) {}
-
-int SimpleFileObject::SimpleFileReadOp::prepare(optional_yield y,
-                                                const DoutPrefixProvider *dpp) {
-  const std::filesystem::path data_path = source->store.object_data_path(
-      source->bucket->get_key(), source->get_key());
-
-  ldpp_dout(dpp, 10) << __func__
-                     << ": TODO bucket_key=" << source->bucket->get_key().name
-                     << " obj_key=" << source->get_key().name
-                     << " path=" << data_path << dendl;
-
-  const auto size = std::filesystem::file_size(data_path);
-  source->set_key(source->get_key());
-  // must set size, otherwise neither read / iterate is called
-  source->set_obj_size(size);
-  return 0;
-}
-
-int SimpleFileObject::SimpleFileReadOp::get_attr(const DoutPrefixProvider *dpp,
-                                                 const char *name,
-                                                 bufferlist &dest,
-                                                 optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO: " << name << dendl;
-
-  if (std::strcmp(name, "user.rgw.acl") == 0) {
-    // TODO support 'user.rgw.acl' to support read_permissions. Return
-    // empty policy since our test user is admin for now.
-    RGWAccessControlPolicy policy;
-    policy.encode(dest);
-    return 0;
-  }
-  return -ENOTSUP;
-}
-
-// sync read
-int SimpleFileObject::SimpleFileReadOp::read(int64_t ofs, int64_t end,
-                                             bufferlist &bl, optional_yield y,
-                                             const DoutPrefixProvider *dpp) {
-  // TODO bounds check, etc.
-  const auto len = end + 1 - ofs;
-  ldpp_dout(dpp, 10) << __func__ << ": TODO offset=" << ofs << " end=" << end
-                     << " len=" << len << dendl;
-
-  const std::filesystem::path object_data_path = source->store.object_data_path(
-      source->bucket->get_key(), source->get_key());
-
-  std::string error;
-  int ret = bl.pread_file(object_data_path.c_str(), ofs, len, &error);
-  if (ret < 0) {
-    ldpp_dout(dpp, 10) << "Failed to read object from file " << object_data_path
-                       << ". Returning EIO" << dendl;
-    return -EIO;
-  }
-  return 0;
-}
-
-// async read
-int SimpleFileObject::SimpleFileReadOp::iterate(const DoutPrefixProvider *dpp,
-                                                int64_t ofs, int64_t end,
-                                                RGWGetDataCB *cb,
-                                                optional_yield y) {
-  // TODO bounds check, etc.
-  const auto len = end + 1 - ofs;
-
-  ldpp_dout(dpp, 10) << __func__ << ": TODO offset=" << ofs << " end=" << end
-                     << " len=" << len << dendl;
-  const std::filesystem::path object_data_path = source->store.object_data_path(
-      source->bucket->get_key(), source->get_key());
-  // TODO chunk the read
-  bufferlist bl;
-  std::string error;
-  int ret = bl.pread_file(object_data_path.c_str(), ofs, len, &error);
-  if (ret < 0) {
-    ldpp_dout(dpp, 10) << "Failed to read object from file " << object_data_path
-                       << ". Returning EIO" << dendl;
-    return -EIO;
-  }
-
-  cb->handle_data(bl, ofs, len);
-  return 0;
-}
-
-SimpleFileObject::SimpleFileDeleteOp::SimpleFileDeleteOp(
-    SimpleFileObject *_source)
-    : source(_source) {}
-
-int SimpleFileObject::SimpleFileDeleteOp::delete_obj(
-    const DoutPrefixProvider *dpp, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::delete_object(const DoutPrefixProvider *dpp,
-                                    optional_yield y, bool prevent_versioning) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::delete_obj_aio(const DoutPrefixProvider *dpp,
-                                     RGWObjState *astate, Completions *aio,
-                                     bool keep_index_consistent,
-                                     optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::copy_object(
-    User *user, req_info *info, const rgw_zone_id &source_zone,
-    rgw::sal::Object *dest_object, rgw::sal::Bucket *dest_bucket,
-    rgw::sal::Bucket *src_bucket, const rgw_placement_rule &dest_placement,
-    ceph::real_time *src_mtime, ceph::real_time *mtime,
-    const ceph::real_time *mod_ptr, const ceph::real_time *unmod_ptr,
-    bool high_precision_time, const char *if_match, const char *if_nomatch,
-    AttrsMod attrs_mod, bool copy_if_newer, Attrs &attrs,
-    RGWObjCategory category, uint64_t olh_epoch,
-    boost::optional<ceph::real_time> delete_at, std::string *version_id,
-    std::string *tag, std::string *etag, void (*progress_cb)(off_t, void *),
-    void *progress_data, const DoutPrefixProvider *dpp, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-/** TODO Create a randomized instance ID for this object */
-void SimpleFileObject::gen_rand_obj_instance_name() {
-  store.ceph_context();
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-  return;
-}
-
-int SimpleFileObject::get_obj_attrs(optional_yield y,
-                                    const DoutPrefixProvider *dpp,
-                                    rgw_obj *target_obj) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileObject::modify_obj_attrs(const char *attr_name,
-                                       bufferlist &attr_val, optional_yield y,
-                                       const DoutPrefixProvider *dpp) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::delete_obj_attrs(const DoutPrefixProvider *dpp,
-                                       const char *attr_name,
-                                       optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-MPSerializer *SimpleFileObject::get_serializer(const DoutPrefixProvider *dpp,
-                                               const std::string &lock_name) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return nullptr;
-}
-
-int SimpleFileObject::transition(Bucket *bucket,
-                                 const rgw_placement_rule &placement_rule,
-                                 const real_time &mtime, uint64_t olh_epoch,
-                                 const DoutPrefixProvider *dpp,
-                                 optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-bool SimpleFileObject::placement_rules_match(rgw_placement_rule &r1,
-                                             rgw_placement_rule &r2) {
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-  return true;
-}
-
-int SimpleFileObject::dump_obj_layout(const DoutPrefixProvider *dpp,
-                                      optional_yield y, Formatter *f) {
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::swift_versioning_restore(bool &restored, /* out */
-                                               const DoutPrefixProvider *dpp) {
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::swift_versioning_copy(const DoutPrefixProvider *dpp,
-                                            optional_yield y) {
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::omap_get_vals(const DoutPrefixProvider *dpp,
-                                    const std::string &marker, uint64_t count,
-                                    std::map<std::string, bufferlist> *m,
-                                    bool *pmore, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileObject::omap_get_all(const DoutPrefixProvider *dpp,
-                                   std::map<std::string, bufferlist> *m,
-                                   optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileObject::omap_get_vals_by_keys(const DoutPrefixProvider *dpp,
-                                            const std::string &oid,
-                                            const std::set<std::string> &keys,
-                                            Attrs *vals) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileObject::omap_set_val_by_key(const DoutPrefixProvider *dpp,
-                                          const std::string &key,
-                                          bufferlist &val, bool must_exist,
-                                          optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-// }}}
-
-// User {{{
-
-int SimpleFileUser::read_attrs(const DoutPrefixProvider *dpp,
-                               optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  /** Read the User attributes from the backing Store */
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::merge_and_store_attrs(const DoutPrefixProvider *dpp,
-                                          Attrs &new_attrs, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  /** Set the attributes in attrs, leaving any other existing attrs set, and
-   * write them to the backing store; a merge operation */
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::read_stats(const DoutPrefixProvider *dpp, optional_yield y,
-                               RGWStorageStats *stats,
-                               ceph::real_time *last_stats_sync,
-                               ceph::real_time *last_stats_update) {
-  /** Read the User stats from the backing Store, synchronous */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::read_stats_async(const DoutPrefixProvider *dpp,
-                                     RGWGetUserStats_CB *cb) {
-  /** Read the User stats from the backing Store, asynchronous */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::complete_flush_stats(const DoutPrefixProvider *dpp,
-                                         optional_yield y) {
-  /** Flush accumulated stat changes for this User to the backing store */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::read_usage(
-    const DoutPrefixProvider *dpp, uint64_t start_epoch, uint64_t end_epoch,
-    uint32_t max_entries, bool *is_truncated, RGWUsageIter &usage_iter,
-    std::map<rgw_user_bucket, rgw_usage_log_entry> &usage) {
-  /** Read detailed usage stats for this User from the backing store */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::trim_usage(const DoutPrefixProvider *dpp,
-                               uint64_t start_epoch, uint64_t end_epoch) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::load_user(const DoutPrefixProvider *dpp, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO (0)" << dendl;
-  return 0;
-}
-
-int SimpleFileUser::store_user(const DoutPrefixProvider *dpp, optional_yield y,
-                               bool exclusive, RGWUserInfo *old_info) {
-  /** Store this User to the backing store */ ldpp_dout(dpp, 10)
-      << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileUser::remove_user(const DoutPrefixProvider *dpp,
-                                optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-static void populate_buckets_from_path(const SimpleFileStore& store, const DoutPrefixProvider *dpp, std::filesystem::path path, BucketList& buckets) {
-  for (auto const &dir_entry : std::filesystem::directory_iterator{path}) {
-    auto bucket =
-        std::unique_ptr<Bucket>(new SimpleFileBucket{dir_entry.path(), store});
-    bucket->load_bucket(dpp, null_yield);
-    buckets.add(std::move(bucket));
-  }
-}
-
-int SimpleFileUser::list_buckets(const DoutPrefixProvider *dpp,
-                                 const std::string &marker,
-                                 const std::string &end_marker, uint64_t max,
-                                 bool need_stats, BucketList &buckets,
-                                 optional_yield y) {
-  // TODO this should list buckets assigned to a user. for now we just get every
-  // bucket
-  populate_buckets_from_path(store, dpp, store.buckets_path(), buckets);
-  ldpp_dout(dpp, 10) << __func__ << ": TODO " << marker << ", " << end_marker
-                     << ", "
-                     << "max=" << max << ", "
-                     << "buckets=" << buckets.get_buckets() << dendl;
-  return 0;
-}
-
-int SimpleFileUser::create_bucket(
-    const DoutPrefixProvider *dpp, const rgw_bucket &b,
-    const std::string &zonegroup_id, rgw_placement_rule &placement_rule,
-    std::string &swift_ver_location, const RGWQuotaInfo *pquota_info,
-    const RGWAccessControlPolicy &policy, Attrs &attrs, RGWBucketInfo &info,
-    obj_version &ep_objv, bool exclusive, bool obj_lock_enabled, bool *existed,
-    req_info &req_info, std::unique_ptr<Bucket> *bucket, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return 0;
-}
-
-// }}}
-
-// Bucket {{{
-
-std::unique_ptr<Object> SimpleFileBucket::get_object(const rgw_obj_key &key) {
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-  /** TODO Get an @a Object belonging to this bucket */
-  return nullptr;
-}
-
-int SimpleFileBucket::list(const DoutPrefixProvider *dpp, ListParams &, int,
-                           ListResults &results, optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": iterating " << objects_path() << dendl;
-  for (auto const &dir_entry :
-       std::filesystem::directory_iterator{objects_path()}) {
-    ldpp_dout(dpp, 10) << __func__ << ": adding object from " << dir_entry
-                       << dendl;
-    if (dir_entry.is_directory()) {
-      JSONParser object_meta_parser;
-      const auto object_meta_path =
-          dir_entry.path() / "rgw_bucket_dir_entry.json";
-      if (!object_meta_parser.parse(object_meta_path.c_str())) {
-        ldpp_dout(dpp, 10) << "Failed to parse object metadata from "
-                           << object_meta_path << ". Skipping" << dendl;
-      }
-      rgw_bucket_dir_entry rgw_dir;
-      rgw_dir.decode_json(&object_meta_parser);
-      results.objs.push_back(rgw_dir);
-    }
-  }
-
-  ldpp_dout(dpp, 10) << __func__ << ": TODO " << dendl;
-  return 0;
-}
-
-int SimpleFileBucket::remove_bucket(const DoutPrefixProvider *dpp,
-                                    bool delete_children,
-                                    bool forward_to_master, req_info *req_info,
-                                    optional_yield y) {
-  /** Remove this bucket from the backing store */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::remove_bucket_bypass_gc(int concurrent_max,
-                                              bool keep_index_consistent,
-                                              optional_yield y,
-                                              const DoutPrefixProvider *dpp) {
-  /** Remove this bucket, bypassing garbage collection.  May be removed */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::load_bucket(const DoutPrefixProvider *dpp,
-                                  optional_yield y, bool get_stats) {
-  std::filesystem::path meta_file_path =
-      bucket_metadata_path("RGWBucketInfo.json");
-  JSONParser bucket_meta_parser;
-  if (!bucket_meta_parser.parse(meta_file_path.c_str())) {
-    ldpp_dout(dpp, 10) << "Failed to parse bucket metadata from "
-                       << meta_file_path << ". Returing EINVAL" << dendl;
-    return -EINVAL;
-  }
-
-  info.decode_json(&bucket_meta_parser);
-  ldpp_dout(dpp, 10) << __func__ << ": TODO " << meta_file_path << dendl;
-  return 0;
-}
-
-int SimpleFileBucket::chown(const DoutPrefixProvider *dpp, User *new_user,
-                            User *old_user, optional_yield y,
-                            const std::string *marker) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-bool SimpleFileBucket::is_owner(User *user) {
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-  return true;
-}
-int SimpleFileBucket::check_empty(const DoutPrefixProvider *dpp,
-                                  optional_yield y) {
-  /** Check in the backing store if this bucket is empty */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::merge_and_store_attrs(const DoutPrefixProvider *dpp,
-                                            Attrs &new_attrs,
-                                            optional_yield y) {
-  /** Set the attributes in attrs, leaving any other existing attrs set, and
-   * write them to the backing store; a merge operation */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-std::unique_ptr<MultipartUpload> SimpleFileBucket::get_multipart_upload(
-    const std::string &oid, std::optional<std::string> upload_id,
-    ACLOwner owner, ceph::real_time mtime) {
-  /** Create a multipart upload in this bucket */
-  return std::unique_ptr<MultipartUpload>();
-}
-
-int SimpleFileBucket::list_multiparts(
-    const DoutPrefixProvider *dpp, const std::string &prefix,
-    std::string &marker, const std::string &delim, const int &max_uploads,
-    std::vector<std::unique_ptr<MultipartUpload>> &uploads,
-    std::map<std::string, bool> *common_prefixes, bool *is_truncated) {
-  /** List multipart uploads currently in this bucket */
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::abort_multiparts(const DoutPrefixProvider *dpp,
-                                       CephContext *cct) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-SimpleFileBucket::SimpleFileBucket(const std::filesystem::path& _path, const SimpleFileStore& _store) : store(_store), path(_path), acls() {
-  ldout(store.ceph_context(), 10) << __func__ << ": TODO" << dendl;
-}
-
-// }}}
-
- // Bucket: Boring Methods {{{
-
-int SimpleFileBucket::try_refresh_info(const DoutPrefixProvider *dpp,
-                                       ceph::real_time *pmtime) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::read_usage(
-    const DoutPrefixProvider *dpp, uint64_t start_epoch, uint64_t end_epoch,
-    uint32_t max_entries, bool *is_truncated, RGWUsageIter &usage_iter,
-    std::map<rgw_user_bucket, rgw_usage_log_entry> &usage) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileBucket::trim_usage(const DoutPrefixProvider *dpp,
-                                 uint64_t start_epoch, uint64_t end_epoch) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::rebuild_index(const DoutPrefixProvider *dpp) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::check_quota(const DoutPrefixProvider *dpp,
-                                  RGWQuotaInfo &user_quota,
-                                  RGWQuotaInfo &bucket_quota, uint64_t obj_size,
-                                  optional_yield y, bool check_size_only) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::read_stats(
-    const DoutPrefixProvider *dpp, int shard_id, std::string *bucket_ver,
-    std::string *master_ver, std::map<RGWObjCategory, RGWStorageStats> &stats,
-    std::string *max_marker, bool *syncstopped) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileBucket::read_stats_async(const DoutPrefixProvider *dpp,
-                                       int shard_id,
-                                       RGWGetBucketStats_CB *ctx) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileBucket::sync_user_stats(const DoutPrefixProvider *dpp,
-                                      optional_yield y) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileBucket::update_container_stats(const DoutPrefixProvider *dpp) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileBucket::check_bucket_shards(const DoutPrefixProvider *dpp) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-int SimpleFileBucket::put_info(const DoutPrefixProvider *dpp, bool exclusive,
-                               ceph::real_time mtime) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-// }}}
-
-// Store > User {{{
-
-std::unique_ptr<User> SimpleFileStore::get_user(const rgw_user &u) {
-  return std::make_unique<SimpleFileUser>(u, *this);
-}
-int SimpleFileStore::get_user_by_access_key(const DoutPrefixProvider *dpp,
-                                            const std::string &key,
-                                            optional_yield y,
-                                            std::unique_ptr<User> *user) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO (returning dummy user)" << dendl;
-  user->reset(new SimpleFileUser(dummy_user, *this));
-  return 0;
-}
-
-int SimpleFileStore::get_user_by_email(const DoutPrefixProvider *dpp,
-                                       const std::string &email,
-                                       optional_yield y,
-                                       std::unique_ptr<User> *user) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  user->reset(new SimpleFileUser(dummy_user, *this));
-  return 0;
-}
-
-int SimpleFileStore::get_user_by_swift(const DoutPrefixProvider *dpp,
-                                       const std::string &user_str,
-                                       optional_yield y,
-                                       std::unique_ptr<User> *user) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-// }}}
-
-// Store > Bucket {{{
-
-int SimpleFileStore::set_buckets_enabled(const DoutPrefixProvider *dpp,
-                                         std::vector<rgw_bucket> &buckets,
-                                         bool enabled) {
-  ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileStore::get_bucket(User *u, const RGWBucketInfo &i,
-                                std::unique_ptr<Bucket> *bucket) {
-  // TODO implement get_bucket by RGWBucketInfo
-  ldout(ctx(), 10) << __func__ << ": TODO get_bucket by RGWBucketInfo" << dendl;
-  return -ENOTSUP;
-}
-
-int SimpleFileStore::get_bucket(const DoutPrefixProvider *dpp, User *u,
-                                const rgw_bucket &b,
-                                std::unique_ptr<Bucket> *result,
-                                optional_yield y) {
-  const auto path = bucket_path(b);
-
-  if (!std::filesystem::exists(path)) {
-    ldpp_dout(dpp, 10) << __func__ << ": bucket "
-                       << " path does not exist: " << path << dendl;
-    return -ENOENT;
-  }
-  auto bucket = make_unique<SimpleFileBucket>(path, *this);
-  const int ret = bucket->load_bucket(dpp, y);
-  if (ret < 0) {
-    return ret;
-  }
-  ldpp_dout(dpp, 10) << __func__ << ": bucket: " << bucket->get_name() << dendl;
-  result->reset(bucket.release());
-  return 0;
-}
-
-int SimpleFileStore::get_bucket(const DoutPrefixProvider *dpp, User *u,
-                                const std::string &tenant,
-                                const std::string &name,
-                                std::unique_ptr<Bucket> *bucket,
-                                optional_yield y) {
-  // TODO implement get_bucket by name
-  ldpp_dout(dpp, 10) << __func__ << ": TODO get_bucket by name" << dendl;
-  return -ENOTSUP;
-}
-
-// }}}
-
 // Lifecycle {{{
 std::unique_ptr<Lifecycle> SimpleFileStore::get_lifecycle(void) {
   ldout(ctx(), 10) << __func__ << ": TODO" << dendl;
@@ -926,76 +272,6 @@ int SimpleFileStore::log_op(const DoutPrefixProvider *dpp, string &oid,
                             bufferlist &bl) {
   ldpp_dout(dpp, 10) << __func__ << ": TODO" << dendl;
   return 0;
-}
-
-// }}}
-
-// Store/Bucket > paths {{{
-
-// buckets_path returns the directory where every subdirectory contains a bucket
-std::filesystem::path SimpleFileStore::buckets_path() const {
-  return data_path / "buckets";
-}
-
-//
-std::filesystem::path SimpleFileStore::users_path() const {
-  return data_path / "users";
-}
-
-// bucket_path returns the directory where bucket metadata is stored
-std::filesystem::path SimpleFileStore::bucket_path(
-    const rgw_bucket &bucket) const {
-  return buckets_path() / bucket.name;
-}
-
-// bucket_path returns the path to a bucket metadata file named metadata_fn
-std::filesystem::path SimpleFileStore::bucket_metadata_path(
-    const rgw_bucket &bucket, const std::string &metadata_fn) const {
-  return bucket_path(bucket) / metadata_fn;
-}
-
-// objects_path returns the directory where all objects of a bucket are stored
-std::filesystem::path SimpleFileStore::objects_path(const rgw_bucket &bucket) const {
-  return bucket_path(bucket) / "objects";
-}
-
-static std::string hash_rgw_obj_key(const rgw_obj_key &obj) {
-  const std::string_view in{obj.name};
-  const auto hash = calc_hash_sha256(in);
-  return hash.to_str();
-}
-
-// object_path returns the directory where metadata of an object is stored
-std::filesystem::path SimpleFileStore::object_path(
-    const rgw_bucket &bucket, const rgw_obj_key &obj) const {
-  return objects_path(bucket) / hash_rgw_obj_key(obj);
-}
-
-// object_data_path returns the path of the object data file
-std::filesystem::path SimpleFileStore::object_data_path(
-    const rgw_bucket &bucket, const rgw_obj_key &obj) const {
-  return object_path(bucket, obj) / "data";
-}
-
-// object_metadata_path returns the path to an object metadata file named
-// metadata_fn
-std::filesystem::path SimpleFileStore::object_metadata_path(
-    const rgw_bucket &bucket, const rgw_obj_key &obj,
-    const std::string &metadata_fn) const {
-  return object_path(bucket, obj) / metadata_fn;
-}
-
-// bucket_path returns the path containing bucket metadata and objects
-std::filesystem::path SimpleFileBucket::bucket_path() const { return path; }
-// bucket_metadata_path returns the path to the metadata file metadata_fn
-std::filesystem::path SimpleFileBucket::bucket_metadata_path(
-    const std::string &metadata_fn) const {
-  return path / metadata_fn;
-}
-// objects_path returns the path to the buckets objects. Each
-// subdirectory points to an object
-std::filesystem::path SimpleFileBucket::objects_path() const {
-  return path / "objects";
 }
 
 // }}}
