+++ src/rgw/rgw_sal_simplefile.h
@@ -25,301 +25,14 @@
 #include "rgw_sal.h"
 #include "rgw_sal_dbstore.h"
 
+#include "store/simplefile/user.h"
+#include "store/simplefile/bucket.h"
+#include "store/simplefile/object.h"
+#include "store/simplefile/zone.h"
+
 namespace rgw::sal {
 
 class SimpleFileStore;
-
-class SimpleFileUser : public User {
- private:
-  const SimpleFileStore& store;
-
- protected:
-  SimpleFileUser(SimpleFileUser&) = default;
-  SimpleFileUser& operator=(const SimpleFileUser&) = default;
-
- public:
-  SimpleFileUser(const rgw_user &_u, const SimpleFileStore& _store) : User(_u), store(_store) {}
-  SimpleFileUser(const RGWUserInfo &_i, const SimpleFileStore& _store) : User(_i), store(_store) {}
-  virtual ~SimpleFileUser() = default;
-
-  virtual std::unique_ptr<User> clone() override {
-    return std::unique_ptr<User>(new SimpleFileUser{*this});
-  }
-  virtual int list_buckets(const DoutPrefixProvider *dpp,
-                           const std::string &marker,
-                           const std::string &end_marker, uint64_t max,
-                           bool need_stats, BucketList &buckets,
-                           optional_yield y) override;
-  virtual int create_bucket(
-      const DoutPrefixProvider *dpp, const rgw_bucket &b,
-      const std::string &zonegroup_id, rgw_placement_rule &placement_rule,
-      std::string &swift_ver_location, const RGWQuotaInfo *pquota_info,
-      const RGWAccessControlPolicy &policy, Attrs &attrs, RGWBucketInfo &info,
-      obj_version &ep_objv, bool exclusive, bool obj_lock_enabled,
-      bool *existed, req_info &req_info, std::unique_ptr<Bucket> *bucket,
-      optional_yield y) override;
-  virtual int read_attrs(const DoutPrefixProvider *dpp,
-                         optional_yield y) override;
-  virtual int merge_and_store_attrs(const DoutPrefixProvider *dpp,
-                                    Attrs &new_attrs,
-                                    optional_yield y) override;
-  virtual int read_stats(const DoutPrefixProvider *dpp, optional_yield y,
-                         RGWStorageStats *stats,
-                         ceph::real_time *last_stats_sync = nullptr,
-                         ceph::real_time *last_stats_update = nullptr) override;
-  virtual int read_stats_async(const DoutPrefixProvider *dpp,
-                               RGWGetUserStats_CB *cb) override;
-  virtual int complete_flush_stats(const DoutPrefixProvider *dpp,
-                                   optional_yield y) override;
-  virtual int read_usage(
-      const DoutPrefixProvider *dpp, uint64_t start_epoch, uint64_t end_epoch,
-      uint32_t max_entries, bool *is_truncated, RGWUsageIter &usage_iter,
-      std::map<rgw_user_bucket, rgw_usage_log_entry> &usage) override;
-  virtual int trim_usage(const DoutPrefixProvider *dpp, uint64_t start_epoch,
-                         uint64_t end_epoch) override;
-  virtual int load_user(const DoutPrefixProvider *dpp,
-                        optional_yield y) override;
-  virtual int store_user(const DoutPrefixProvider *dpp, optional_yield y,
-                         bool exclusive,
-                         RGWUserInfo *old_info = nullptr) override;
-  virtual int remove_user(const DoutPrefixProvider *dpp,
-                          optional_yield y) override;
-};
-
-class SimpleFileBucket : public Bucket {
- private:
-  const SimpleFileStore& store;
-  const std::filesystem::path path;
-  RGWAccessControlPolicy acls;
- protected:
-  SimpleFileBucket(const SimpleFileBucket&) = default;
-
- public:
-  SimpleFileBucket(const std::filesystem::path& _path, const SimpleFileStore& _store);
-  SimpleFileBucket& operator=(const SimpleFileBucket&) = delete;
-
-  std::filesystem::path bucket_path() const;
-  std::filesystem::path bucket_metadata_path(const std::string& metadata_fn) const;
-  std::filesystem::path objects_path() const;
-
-  virtual std::unique_ptr<Bucket> clone() override {
-    return std::unique_ptr<Bucket>(new SimpleFileBucket{*this});
-  }
-
-  virtual std::unique_ptr<Object> get_object(const rgw_obj_key &key) override;
-  virtual int list(const DoutPrefixProvider *dpp, ListParams &, int,
-                   ListResults &, optional_yield y) override;
-  virtual int remove_bucket(const DoutPrefixProvider *dpp, bool delete_children,
-                            bool forward_to_master, req_info *req_info,
-                            optional_yield y) override;
-  virtual int remove_bucket_bypass_gc(int concurrent_max,
-                                      bool keep_index_consistent,
-                                      optional_yield y,
-                                      const DoutPrefixProvider *dpp) override;
-  virtual RGWAccessControlPolicy &get_acl(void) override { return acls; }
-
-  virtual int set_acl(const DoutPrefixProvider *dpp,
-                      RGWAccessControlPolicy &acl, optional_yield y) override {
-    acls = acl;
-    return 0;
-  }
-  virtual int load_bucket(const DoutPrefixProvider *dpp, optional_yield y,
-                          bool get_stats = false) override;
-  virtual int read_stats(const DoutPrefixProvider *dpp, int shard_id,
-                         std::string *bucket_ver, std::string *master_ver,
-                         std::map<RGWObjCategory, RGWStorageStats> &stats,
-                         std::string *max_marker = nullptr,
-                         bool *syncstopped = nullptr) override;
-  virtual int read_stats_async(const DoutPrefixProvider *dpp, int shard_id,
-                               RGWGetBucketStats_CB *ctx) override;
-  virtual int sync_user_stats(const DoutPrefixProvider *dpp,
-                              optional_yield y) override;
-  virtual int update_container_stats(const DoutPrefixProvider *dpp) override;
-  virtual int check_bucket_shards(const DoutPrefixProvider *dpp) override;
-  virtual int chown(const DoutPrefixProvider *dpp, User *new_user,
-                    User *old_user, optional_yield y,
-                    const std::string *marker = nullptr) override;
-  virtual int put_info(const DoutPrefixProvider *dpp, bool exclusive,
-                       ceph::real_time mtime) override;
-  virtual bool is_owner(User *user) override;
-  virtual int check_empty(const DoutPrefixProvider *dpp,
-                          optional_yield y) override;
-  virtual int check_quota(const DoutPrefixProvider *dpp,
-                          RGWQuotaInfo &user_quota, RGWQuotaInfo &bucket_quota,
-                          uint64_t obj_size, optional_yield y,
-                          bool check_size_only = false) override;
-  virtual int merge_and_store_attrs(const DoutPrefixProvider *dpp,
-                                    Attrs &new_attrs,
-                                    optional_yield y) override;
-  virtual int try_refresh_info(const DoutPrefixProvider *dpp,
-                               ceph::real_time *pmtime) override;
-  virtual int read_usage(
-      const DoutPrefixProvider *dpp, uint64_t start_epoch, uint64_t end_epoch,
-      uint32_t max_entries, bool *is_truncated, RGWUsageIter &usage_iter,
-      std::map<rgw_user_bucket, rgw_usage_log_entry> &usage) override;
-  virtual int trim_usage(const DoutPrefixProvider *dpp, uint64_t start_epoch,
-                         uint64_t end_epoch) override;
-  virtual int rebuild_index(const DoutPrefixProvider *dpp) override;
-  virtual std::unique_ptr<MultipartUpload> get_multipart_upload(
-      const std::string &oid,
-      std::optional<std::string> upload_id = std::nullopt, ACLOwner owner = {},
-      ceph::real_time mtime = real_clock::now()) override;
-  virtual int list_multiparts(
-      const DoutPrefixProvider *dpp, const std::string &prefix,
-      std::string &marker, const std::string &delim, const int &max_uploads,
-      std::vector<std::unique_ptr<MultipartUpload>> &uploads,
-      std::map<std::string, bool> *common_prefixes,
-      bool *is_truncated) override;
-  virtual int abort_multiparts(const DoutPrefixProvider *dpp,
-                               CephContext *cct) override;
-
-  // maybe removed from api..
-  virtual int remove_objs_from_index(
-      const DoutPrefixProvider *dpp,
-      std::list<rgw_obj_index_key> &objs_to_unlink) override {
-    return -ENOTSUP;
-  }
-  virtual int check_index(
-      const DoutPrefixProvider *dpp,
-      std::map<RGWObjCategory, RGWStorageStats> &existing_stats,
-      std::map<RGWObjCategory, RGWStorageStats> &calculated_stats) override {
-    return -ENOTSUP;
-  }
-  virtual int set_tag_timeout(const DoutPrefixProvider *dpp,
-                              uint64_t timeout) override {
-    return 0;
-  }
-  virtual int purge_instance(const DoutPrefixProvider *dpp) override {
-    return -ENOTSUP;
-  }
-};
-
-class SimpleFileObject : public Object {
- private:
-  const SimpleFileStore& store;
-  RGWAccessControlPolicy acls;
- protected:
-  SimpleFileObject(SimpleFileObject&) = default;
-
- public:
-  struct SimpleFileReadOp : public ReadOp {
-   private:
-    SimpleFileObject *source;
-    RGWObjectCtx *rctx;
-
-   public:
-    SimpleFileReadOp(SimpleFileObject *_source, RGWObjectCtx *_rctx);
-
-    virtual int prepare(optional_yield y,
-                        const DoutPrefixProvider *dpp) override;
-    virtual int read(int64_t ofs, int64_t end, bufferlist &bl, optional_yield y,
-                     const DoutPrefixProvider *dpp) override;
-    virtual int iterate(const DoutPrefixProvider *dpp, int64_t ofs, int64_t end,
-                        RGWGetDataCB *cb, optional_yield y) override;
-    virtual int get_attr(const DoutPrefixProvider *dpp, const char *name,
-                         bufferlist &dest, optional_yield y) override;
-  };
-  struct SimpleFileDeleteOp : public DeleteOp {
-   private:
-    SimpleFileObject *source;
-
-   public:
-    SimpleFileDeleteOp(SimpleFileObject *_source);
-    virtual int delete_obj(const DoutPrefixProvider *dpp,
-                           optional_yield y) override;
-  };
-  SimpleFileObject& operator=(const SimpleFileObject&) = delete;
-
-  SimpleFileObject(const SimpleFileStore& _st, const rgw_obj_key &_k)
-      : Object(_k), store(_st) {}
-  SimpleFileObject(const SimpleFileStore& _st, const rgw_obj_key &_k, Bucket *_b)
-      : Object(_k, _b), store(_st) {}
-
-  virtual std::unique_ptr<Object> clone() override {
-    return std::unique_ptr<Object>(new SimpleFileObject{*this});
-  }
-
-  virtual int delete_object(const DoutPrefixProvider *dpp, optional_yield y,
-                            bool prevent_versioning = false) override;
-  virtual int delete_obj_aio(const DoutPrefixProvider *dpp, RGWObjState *astate,
-                             Completions *aio, bool keep_index_consistent,
-                             optional_yield y) override;
-  virtual int copy_object(
-      User *user, req_info *info, const rgw_zone_id &source_zone,
-      rgw::sal::Object *dest_object, rgw::sal::Bucket *dest_bucket,
-      rgw::sal::Bucket *src_bucket, const rgw_placement_rule &dest_placement,
-      ceph::real_time *src_mtime, ceph::real_time *mtime,
-      const ceph::real_time *mod_ptr, const ceph::real_time *unmod_ptr,
-      bool high_precision_time, const char *if_match, const char *if_nomatch,
-      AttrsMod attrs_mod, bool copy_if_newer, Attrs &attrs,
-      RGWObjCategory category, uint64_t olh_epoch,
-      boost::optional<ceph::real_time> delete_at, std::string *version_id,
-      std::string *tag, std::string *etag, void (*progress_cb)(off_t, void *),
-      void *progress_data, const DoutPrefixProvider *dpp,
-      optional_yield y) override;
-
-  virtual RGWAccessControlPolicy &get_acl(void) override { return acls; }
-  virtual int set_acl(const RGWAccessControlPolicy &acl) override {
-    acls = acl;
-    return 0;
-  }
-
-  virtual int get_obj_attrs(optional_yield y, const DoutPrefixProvider *dpp,
-                            rgw_obj *target_obj = NULL) override;
-  virtual int modify_obj_attrs(const char *attr_name, bufferlist &attr_val,
-                               optional_yield y,
-                               const DoutPrefixProvider *dpp) override;
-  virtual int delete_obj_attrs(const DoutPrefixProvider *dpp,
-                               const char *attr_name,
-                               optional_yield y) override;
-  virtual bool is_expired() override { return false; }
-  virtual void gen_rand_obj_instance_name() override;
-  virtual MPSerializer *get_serializer(const DoutPrefixProvider *dpp,
-                                       const std::string &lock_name) override;
-
-  virtual int transition(Bucket *bucket,
-                         const rgw_placement_rule &placement_rule,
-                         const real_time &mtime, uint64_t olh_epoch,
-                         const DoutPrefixProvider *dpp,
-                         optional_yield y) override;
-  virtual bool placement_rules_match(rgw_placement_rule &r1,
-                                     rgw_placement_rule &r2) override;
-  virtual int dump_obj_layout(const DoutPrefixProvider *dpp, optional_yield y,
-                              Formatter *f) override;
-  virtual int swift_versioning_restore(bool &restored, /* out */
-                                       const DoutPrefixProvider *dpp) override;
-  virtual int swift_versioning_copy(const DoutPrefixProvider *dpp,
-                                    optional_yield y) override;
-  virtual std::unique_ptr<ReadOp> get_read_op() override {
-    return std::make_unique<SimpleFileObject::SimpleFileReadOp>(this, nullptr);
-  }
-  virtual std::unique_ptr<DeleteOp> get_delete_op() override {
-    return std::make_unique<SimpleFileObject::SimpleFileDeleteOp>(this);
-  }
-  virtual int omap_get_vals(const DoutPrefixProvider *dpp,
-                            const std::string &marker, uint64_t count,
-                            std::map<std::string, bufferlist> *m, bool *pmore,
-                            optional_yield y) override;
-  virtual int omap_get_all(const DoutPrefixProvider *dpp,
-                           std::map<std::string, bufferlist> *m,
-                           optional_yield y) override;
-  virtual int omap_get_vals_by_keys(const DoutPrefixProvider *dpp,
-                                    const std::string &oid,
-                                    const std::set<std::string> &keys,
-                                    Attrs *vals) override;
-  virtual int omap_set_val_by_key(const DoutPrefixProvider *dpp,
-                                  const std::string &key, bufferlist &val,
-                                  bool must_exist, optional_yield y) override;
-  // will be removed in the future..
-  virtual int get_obj_state(const DoutPrefixProvider *dpp, RGWObjState **state,
-                            optional_yield y, bool follow_olh = true) override {
-    return 0;
-  }
-  virtual int set_obj_attrs(const DoutPrefixProvider *dpp, Attrs *setattrs,
-                            Attrs *delattrs, optional_yield y) override {
-    return 0;
-  }
-};
 
 class UnsupportedLuaScriptManager : public LuaScriptManager {
  public:
@@ -342,34 +55,6 @@ class UnsupportedLuaScriptManager : public LuaScriptManager {
   }
 };
 
-class SimpleFileZone : public Zone {
- protected:
-  SimpleFileStore *store;
-  RGWRealm *realm{nullptr};
-  RGWZoneGroup *zonegroup{nullptr};
-  RGWZone *zone_public_config{nullptr};
-  RGWZoneParams *zone_params{nullptr};
-  RGWPeriod *current_period{nullptr};
-  rgw_zone_id cur_zone_id;
-
- public:
-  SimpleFileZone(const SimpleFileZone&) = delete;
-  SimpleFileZone& operator= (const SimpleFileZone&) = delete;
-  SimpleFileZone(SimpleFileStore *_store);
-  ~SimpleFileZone() = default;
-
-  virtual const RGWZoneGroup &get_zonegroup() override;
-  virtual int get_zonegroup(const std::string &id,
-                            RGWZoneGroup &zonegroup) override;
-  virtual const RGWZoneParams &get_params() override;
-  virtual const rgw_zone_id &get_id() override;
-  virtual const RGWRealm &get_realm() override;
-  virtual const std::string &get_name() const override;
-  virtual bool is_writeable() override;
-  virtual bool get_redirect_endpoint(std::string *endpoint) override;
-  virtual bool has_zonegroup_api(const std::string &api) const override;
-  virtual const std::string &get_current_period_id() override;
-};
 
 class SimpleFileStore : public Store {
  private:
